requirements.txt
alembic==1.4.2
altgraph==0.17
aniso8601==8.0.0
APScheduler==3.6.3
arrow==0.15.6
atoma==0.0.17
attrs==19.3.0
auto-py-to-exe==2.7.5
Automat==20.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.0
bleach==3.1.5
blinker==1.4
Bootstrap-Flask==1.3.2
bottle==0.12.18
bottle-websocket==0.2.9
breadability==0.1.20
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
colorgram.py==1.2.0
colorhash==1.0.2
colorthief==0.2.1
configparser==5.0.0
constantly==15.1.0
crayons==0.4.0
cryptography==2.8
cssselect==1.1.0
defusedxml==0.6.0
docopt==0.6.2
dominate==2.4.0
Eel==0.12.4
elasticsearch==7.7.1
Faker==4.0.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap4==4.0.2
Flask-Caching==1.8.0
Flask-Cors==3.0.8
flask-crontab==0.1.2
Flask-DebugToolbar==0.11.0
Flask-FontAwesome==0.1.4
Flask-Googletrans==0.10
Flask-Limiter==1.2.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Matomo==1.2.0
Flask-Migrate==2.5.3
Flask-MonitoringDashboard==3.0.9
flask-msearch==0.2.9
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
future==0.18.2
get==2019.4.13
gevent==20.5.0
gevent-websocket==0.10.1
googletrans==2.4.0
greenlet==0.4.15
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.15.1
jsonschema==3.2.0
limits==1.5.1
lxml==4.4.1
Mako==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
marshmallow==3.6.1
nltk==3.5
numpy==1.18.2
packaging==20.4
pandas==1.0.2
parsel==1.5.2
pefile==2019.4.18
Pillow==7.2.0
post==2019.4.13
Protego==0.1.16
psutil==5.7.0
psycopg2-binary==2.8.5
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==19.8.18
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyInstaller==3.6
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
pyrsistent==0.16.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
pywin32-ctypes==0.2.0
query-string==2019.4.13
queuelib==1.5.0
rake-nltk==1.0.4
readability-lxml==0.8.1
redis==3.5.0
regex==2020.5.14
request==2019.4.13
requests==2.23.0
rq==1.3.0
rq-dashboard==0.6.1
Scrapy==2.0.0
selenium==3.141.0
sentry-sdk==0.14.3
service-identity==18.1.0
shortuuid==1.0.1
six==1.14.0
soupsieve==1.9.3
SQLAlchemy==1.3.15
sumy==0.8.1
text-unidecode==1.3
tinydb==3.14.1
tqdm==4.46.0
twilio==6.39.0
Twisted==19.10.0
tzlocal==2.0.0
urllib3==1.25.8
visitor==0.1.3
w3lib==1.21.0
webdriver-manager==3.2.2
webencodings==0.5.1
Werkzeug==0.16.1
whichcraft==0.6.1
WTForms==2.2.1
xlrd==1.2.0
XlsxWriter==1.2.1
zipp==3.1.0
zope.interface==4.7.2
